requirements.txt
squidpy[interactive]
harmony-pytorch
jupyter-tabnine
python-lsp-server[all]
mlflow[extras]
fsspec==2022.1.0
gcsfs==2022.1.0
leidenalg==0.10.1
observable-jupyter==0.1.12
clustergrammer2==0.18.0
loompy==3.0.6
pygeos==0.12.0
geopandas==0.10.2
numpy
scipy
scanpy==1.8.2
pandas
networkx
scikit-learn
matplotlib<3.7
geovoronoi
torch >= 1.13.0
torch_geometric >= 2.2.0
torch_sparse
torch_scatter
openpyxl
rtree
panel
kaleido
torchinfo
opencv-python
dash-cytoscape